Restructure asset directory & update image path
